# student-assignment-api/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Student Assignment Tracker API",
        "description": "REST API for managing students and assignments with GitHub OAuth authentication.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://cse341-projects-p64y.onrender.com",
            "description": "Production server"
        },
        {
            "url": "http://localhost:8080",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/auth/github/callback": {
            "get": {
                "summary": "GitHub OAuth callback",
                "description": "Handles the callback from GitHub after successful authentication.",
                "responses": {
                    "200": {
                        "description": "Authentication successful"
                    },
                    "401": {
                        "description": "Authentication failed"
                    }
                }
            }
        },
        "/auth/status": {
            "get": {
                "summary": "Check authentication status",
                "description": "Returns the current user's authentication status.",
                "responses": {
                    "200": {
                        "description": "User is authenticated"
                    },
                    "401": {
                        "description": "User is not authenticated"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sessionAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "connect.sid",
                "description": "Authentication is provided through the Express session created after GitHub OAuth login."
            }
        },
...
    }
}